+++ views/global/html.php
@@ -1,6 +1,7 @@
 <?php
 
 use humhub\modules\custom_pages\widgets\CustomPageInlineStyle;
+use humhub\modules\file\widgets\ShowFiles;
 use yii\helpers\Html;
 
 /** @var $page \humhub\modules\custom_pages\models\Page */